media-source/MediaSource-sourceBuffer.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Initial state of MediaSource#sourceBuffers</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var mediaSource = new MediaSource();
    assert_equals(mediaSource.sourceBuffers.length, 0);
}, 'MediaSource#sourceBuffers : Initial state');
</script>
</body>
</html>
